Unify TailwindCSS dependency for the main page and subject pages
Previously, due to the dependency on tailwindcss typography, the individual pages (e.g. chemistry, english, etc.) had to use a separate tailwind.config.js which built a separate CSS file exclusively for the pages.  This resulted in duplicated CSS and a very messy dependency tree due to the links from each page to another often breaking on changes. Now, there is only 1 CSS file which includes all of the tailwindcss-typography styles, which greatly reduces this issue.